+++ hikka/inline/core.py
@@ -1,6 +1,6 @@
 """Inline buttons, galleries and other Telegram-Bot-API stuff"""
 
-# ©️ Dan Gazizullin, 2021-2022
+# ©️ Dan Gazizullin, 2021-2023
 # This file is a part of Hikka Userbot
 # 🌐 https://github.com/hikariatama/Hikka
 # You can redistribute it and/or modify it under the terms of the GNU AGPLv3